grpc-netty: fix flaky SingleRequestOrResponseApiTest (#3064)
Motivation:

We have a flaky test. It turns out that it should always fail, but we
get lucky and short circuit a `Publisher.cancel()` call the vast
majority of the time. However, fixing that is going to be relatively
hard/ugly.

Modifications:

Because we're guarding against bad gRPC implementations, not app usage,
and the server behavior is correct (it doesn't process the message in
any case), the best thing to do is relax the test requirements and
accept the CANCEL result for the cases where it wins the race. This
result is slightly less debuggable from the client perspective, but the
behavior is acceptable.

Result:

One less flaky test.