http-netty: Fix flaky FullDuplexAndSequentialModeTest
Motivation:

This test suite has been flaky for a while. As detailed in issue #1894,
this is because under some conditions the test running thread ends up
waiting on a latch that it is actually responsible to release.

Modifications:

Make sure the publishing of the InputStreamPublisher happens in on a
different thread so we don't end up in deadlock.

Fixes #1894