http-netty: Add phantom reference version of watchdog filters
Motivation

The current watchdog filters are not exact: they don't account for
responses that get drained later than the cleaner expects which can
result in false positives. Because of this they can't proactively
drain the responses.

Modifications

Add a PhantomReference variant of the watchdog filters. This filter
uses a PhantomReference based implementation to be certain the
response has been abandoned. Because we can now be certain, we can
also proactively drain the response.