http-netty: Re-enable ClientEfectivenessStrategyTest.clientStrategy
Motivation:

We had previously disabled the clientStrategy test because it's flaky.
According to the notes in the issue (#2245) it is only in one part,
and that is in the Send stage which appears to be offloaded when it
shouldn't be.

Modifications:

Re-enable the test and prune those errors, and log them.