traffic-resilience-http: Fix flaky testStopAcceptingConnections() test (#3125)
Motivation:

The `TrafficResilienceHttpServiceFilterTest.testStopAcceptingConnections(..)` test is flaky and sometimes will find that a connection times out before it was expected to. These additional connections are made to account for intrinsic races as well as kernel connection backlog behavior which is really hard to control.

Modifications:

The timeout behavior is the desired result, it just happens before we expect it to. So, instead of trying to make exactly two connections that we expect to succeed followed by one that should fail, we instead can just perform a 3 iteration loop and demand that we eventually stop accepting connections (presuming it's not a dry-run), which is the desired behavior.

Closes #3076.