loadbalancer-experimental: rename ConnectionPoolStrategy -> ConnectionSelector (#3130)
Motivation:

The ConnectionPoolStrategy carries a lot of parallels with the HostSelector
abstraction, in that it's goal is to select a connection. However, the names
are dissimilar and 'Strategy' doesn't have as strong an implication of
being capable of action.

Modifications:

Rename ConnectionPoolStrategy to ConnectionSelector to make its role more
clear.