[grpc-protoc] Add an option to generate default service methods (#3110)
### Motivation

When evolving a service definition that has multiple implementations it is ideal not to break these implementations. Therefore it would be nice if the protoc generator was able to generate default implementations on the service interface to ensure implementations always conform.

### Modifications

Add an option, `defaultServiceMethods` to the grpc-protoc stub compiler to generate these default implementations. The result is that service interfaces will provide a default, throwing `UNIMPLEMENTED` exception, implementation of all service RPC interfaces that they implement.