loadbalancer-experimental: Add default policies
Motivation:

Users currently need to understand a lot about the different policies
in order to make a good decision.

Modifications:

Add default connection pool and load balancing policies. This makes it
easy for users to pick the recommended default. This also gives us as
maintainers a place to change the recommended default.