Preserve exception type on BlockingIterable#forEach variants (#3154)
The previous refactoring wrapped all non-runtime exception into
a runtime exception to adhere to the unchecked singature of
the method. Turns out, some code on top has been relying on the
actual exception (i.e. an IOException) so wrapping, while correct,
is not backwards compatible and in this case for us the latter
is more important.