Improve discovery of (and fix some) netty-releated test buffer leaks
This changeset improves disovery of netty-related buffer leaks if
the --warn flag is enabled on the command line when running the
gradle tests. Paranoid is enabled all the time now, but the overhead
of doing so is very minimal and does not need to be enabled conditionally.
--warn is needed since only in that mode stdout is visible on the
command line.

Running it showed a couple of leaks in tests where two of them are
already fixed as part of this commit.