scripts/release.sh: ensure tree is clean before releasing
Motivation:

As part of the release process our script makes some changes to versions
etc and then everything in the branch is committed. This is dangerous if
there are uncommitted changes present.

Modifications:

As part of the release.sh script check if the tree is dirty and error
out if so.