concurrent-api: cleanup AsyncContext operations
Motivation:

Perhaps the main goal of the AsyncContext is to allow a way to bundle up
thread-local (ish) state and restore it when an async operation occurs.
However, restore operation is essentially inlined in all the context
operations which makes the mechanism more difficult to understand than
it needs to be.

Modifications:

Add a few new methods to the AsyncContextProvider

- saveContext() the goal of this method is to package up context for
  propagation to the next continuation. This stands in contrast to the
  existing context() method which is used for reading the async ContextMap.
- attachContext(ContextMap) this method is used to restore the local state
  for temporary use. It returns a new Scope type, which then reverts
  this local environment to what it was before the restore.

Result:

Cleaner async context operations.