concurrent-api: shave some allocations from new Scope (#3185)
Motivation:

In the allocation profile we see anonymous instances of the new `Scope` being
created. While not a big deal, we can do better.

Modifications:

We can shave these off of the default implementation by making `CopyOnWriteContextMap`,
and really any context map, an instance of `Scope`. The `close()` method just needs
to then ask the current provider to set the context to itself.

Result:

Less allocations.