http-utils: Fix bad logging when request and response are present (#3191)
Motivation:

We use the var-args logger in LoggingHttpLifecycleObserver and add the
exception regardless of if one is generated which means when there is not
an exception we put a `null` as the last argument. The logger can't tell
this was supposed to be an exception so it gets angry because it found
one more argument (a null) than it expected based on the format string.

Modifications:

Don't add the exception if it's null.