loadbalancer: fix flaky test (#3195)
Motivation:

We have a flaky test. This is because we're able to get a connection
but only transiently. This happens because the DefaultHost is willing
to add a connection to an expired host without connections, which is a
state that is about to be closed.

Modifications:

Don't add the connection to the host if the host is is the transient
expired state with zero connections.

Fixes #2677