concurrent-api: defer logging during CapturedContextProviders initialization
Motivation:

The MDC utilities depend on the AsyncContext for storing properties.
This can create a cyclical dependency since we try to log the
loaded providers as part of the initialization.

Modifications:

Submit the logging to a different thread so that it will happen 'later'
after the class initialization is complete.