http-netty: Unify how we drain the request body if it's unused
Motivation:

We have some special logic for request draining in the http server
but it's overly complicated.

Modifications:

Consolidate on the signalling that we already expect to happen in
the case the application consumes the request.

This makes the pattern simpler, and one that can be lifted out
into a filter later.