`DefaultNettyConnection`: share AsyncContext state on write (#3218)
Motivation:

Even though we pass a `Publisher` to `NettyConnection.write(...)` as
another async source, we always use it in a way that expects a shared
AsyncContext state between what we write and returned `Completable`.
However, we can not apply `shareContextOnSubscribe()` on the passed
`Publisher` because internally `write(...)` operation also applies
`FlushSubscriber`.

Modifications:

- Add `shareContextOnSubscribe()` before subscribing to the write
`Publisher`.
- Add a test to assert expected context propagation.

Result:

`AsyncContext` state is shared between the `Publisher` we write and
returned `Completable`.