http-netty: fix flaky ExpectContinueTest.retryExpectationFailed
Motivation:

This test is flaky because of a race condition where we let the retry
potentially happen before we assert that the request body hasn't been
subscribed to. On the retry the entire request is sent without the
expect: continue header so it should all be sent, and thus we expect
the body to be subscribed to.

Modifications:

Don't let the first response continue until after the assertion. This
should still test that the first expect: continue request doesn't
subscribe to the body until it knows it's safe to proceed.

Result:

One less flaky test.

Fixes #2638