http-netty: Properly drain request body if response body fails (#3224)
Motivation:

We use the request and response bodies heavily to signal the end of http dispatch lifecycles. This means we really need them to be properly drained. The NettyHttpServer has an auto-drain feature aimed at making sure request bodies are disposed of, but it is not effective if the response body results in error.

Modification:

Change the request draining to happen unconditionally regardless of if the write succeeds or not.