opentelemetry-http: Set context correctly for the server request body (#3212)
Motivation:

We need to also consider the request body when tracking signals and setting context.

Modifications:

Properly track the response. Note that we must set the auto-draining to happen
later that tracing in the pipeline to get the correct span info on all request body
callbacks, but this isn't a new problem introduced in this PR.