log4j-mdc: Use CapturedContext mechanism for MDC to get copy behavior
Motivation:

The MDC context currently lives in the ServiceTalk AsyncContext storage
system. The AsyncContext itself has a notion of copy vs sharing on
subscribe, but the MDC context, being stored as a value in that map,
does not inherit that behavior. That makes for strange behavior when
the MDC context ends up crossing boundaries unexpectedly.

Modifications:

Use the CapturedContext mechanism which gives a way to provide a notion
of copy vs share.

Result:

More consistent context behavior.