http-netty: pass a useful id to the load-balancer construction process
Motivation:

When we create the default load balancer for our HttpClientBuilder
we pass the class name of the DefaultHttpLoadBalancerFactory. This isn't
helpful for distinguishing between different load balancers for different
endpoints which makes it really hard for providers to operate on a single
load balancers client, instead it can only operate on all clients.

Modifications:

Instead of passing in the class name as the ID, we pass in the address
in string form.

Result:

Providers can operate on individual clients.