transport-netty-internal: properly order observer events if connection establishment fails (#3238)
Motivation:

When establishing a TCP connection, we add a listener to the `closeFuture()`
of the channel to let us know when close events happen. This works fine
until we consider the tcp connect: it turns out that the connect future
listeners will typically be added before the attachment of channel
closure-on-failure. This cases the Subscriber.onError pathway to be
triggered before the ConnectionObserver.connectionClosed(failed)
pathway because the former has its subscriber added to the promise
first.

Modifications:

Refactor the code to consider the connect pathway and add an early
listener to the promise that will trigger the connectionClosed(..)
callback eagerly. If the connection succeeds we then attach a listener
to the `closeFuture()` as usual.