Reduce PR quality matrix to just JDK 8 and 11
Motivation:

We have a lot of jobs in our build and run each of them against
every JDK version we want to support. The for quality checks,
this is unnecessary because we only emit artifacts for JDK 8 and 11.

Modifications:

Reduce the PR quality checks to only run for JDK 8 and 11.

Result:

Less build work.