Validate generated `pom.xml` file for all published libraries (#3262)
Motivation:

We had the same problem twice (see #3078 and #3261), when produced `pom.xml` file for some of our modules was invalid from Maven perspective, and it produced the following warnings:
```
[WARNING] The POM for io.servicetalk:servicetalk-log4j2-mdc:jar:0.42.56 is invalid, transitive dependencies (if any) will not be available: 1 problem was encountered while building the effective model for io.servicetalk:servicetalk-log4j2-mdc:0.42.56
[ERROR] 'dependencies.dependency.version' for org.apache.logging.log4j:log4j-core:jar is missing. @
```
For some reason, Gradle can still resolve dependencies with a missed version, even if their BOM is also missing, but Maven can not.

Modifications:

1. Enhance `MavenPublication` task to validate the generated `pom.xml` and fail if any dependency does not have a version or corresponding BOM.
2. Make `generatePomFileForMavenJavaPublication` part of `quality` task to enforce this validation on every PR.

Result:

Hopefully, this issue won't happen again. Tested that it failed the build before #3261 was merged.