opentelemetry-http: fox url extraction
Motivation:

Our url extraction is wrong in that it doesn't consider possible
absolute-form request targets. We also don't allow users to set
a host and port in the common case that it's using origin-form
request targets and we can't extract a host name or port.

Modifications:

Fix the extraction and make it possible to provide a host and port.