tcp-netty-internal: Make sure context information is available to ConnectionObserver (#3257)
Motivation:

We sometimes make new connections as part of the request path
and that means it's important to make sure the context information
is available for the observers so we can instrument them.

Modifications:

Capture the context present when the connection is created. We
then make sure it's there for the right methods in the
`ConnectionObserver`. Note that the `ConnectionObserver` has
a lifecycle that lives beyond a request, and thus a single context,
so we drop that context information after the initial connection
establishment (or failure).