Tighten usage of Vary HTTP header
Motivation:
The current behavior around the Vary HTTP header has two issues associated:

 - It is a response-only header, so the client should not send it when
   sending a compressed payload.
 - The Vary header describes the parts of the request message that influenced
   the content of the response it occurs in. This in practice means for a server
   to return Accept-Encoding, but we return Content-Encoding which is not correct.

Modifications:
This changeset first makes sure that the client actually never sends the vary
header with its request, even if the payload is being compressed. Then, if the
vary header is set by the server side it is not set to Content-Encoding but
rather to Accept-Encoding, which influenced the compression decision on the
response from the incoming request.

Result:
Vary header behavior which aligns better with HTTP protocol expectations.