concurrent-api: improve context wrapping (#3276)
Motivation:

We do a lot of wrapping, and a lot of work deciding how
to wrap. This can be simplified in a few ways.

Modifications:

We know we don't need to wrap twice: the inner wrapper
will just override the outer wrapper. However, we have
both the Cancellable and the Subscriber to worry about
and they're handled by three different types.

Instead, we can just have one type and its job is to
do it all, conditionally if it applies, of course. So,
we shift all the work to ContextPreservingCompletableSubscriber
and make it conditionally wrap depending what context is
present.

Note: since we had to remove the `ContextMap saved` field
in the wrappers, we're going to remove the rest as well.

Result:

- More efficient code: we now don't double wrap in a couple
  cases.
- More readable code: we only have one wrapper type which
  really simplifies diffing through wrappers to see whats
  going on.