opentelemetry-http: defer client attribute extraction
Motivation:

Current client attribute extraction happens on start
of the filter which is usually at the front of the
pipeline. This doesn't let client filters operate on
the request which can often populate important data
such as the host header, used for attribute extraction.

Modifications:

Defer client attribute extraction to the `onEnd` phase.
Since nothing in there is time sensitive it doesn't
make a difference other than the request is as it
is intended.