Migrate to central portal (#3289)
Motivation:

The OSSRH endpoints are deprecated and being shut off at
the end of the month.

Modifications:

Switch to the central portal endpoints for publishing.
This endpoint is still compatible with the gradle plugins,
it just requires changing the actual endpoint and switching
credentials.

Result:

We can hopefully still publish.