`TransportObserver`: let users easily skip uninterested callbacks
Motivation:

`TransportObserver` and all other observers related to it, like
`ConnectionObserver`, etc. have quite a lot of methods but users are not
interested in all of them at the same time. However, the current
interface definition forces users to maintain a lot of boilerplate code.

Modifications:
- Add default implementations for all `ConnectionObserver` methods and
for methods of other related observers.
- Remove boilerplate from existing `TransportObserver` implementations.
- `DefaultNettyConnection`: skip wiring `ReadObserver` if it's noop.

Result:

`TransportObserver` implementations that are interested only in subset
of callbacks will be able to remove boilerplate code.