`[Http|Grpc]LifecycleObserver`: let users skip uninterested callbacks
#### Motivation:

`[Http|Grpc]LifecycleObserver` and all other observers related to it,
like `HttpExchangeObserver`, etc. have quite a lot of methods but users
are not interested in all of them at the same time. However, the current
interface definition forces users to maintain a lot of boilerplate code.

Modifications:
- Add default implementations for all `HttpExchangeObserver` methods and
for methods of other related observers.
- Remove boilerplate from existing observer implementations.

Result:

`[Http|Grpc]LifecycleObserver` implementations that are interested only
in subset of callbacks will be able to remove boilerplate code.

Similar to #3297.