Make all `DefaultHttpServerBuilder` methods `final` except one we need
Motivation:

`DefaultHttpServerBuilder` is a pkg-private class that was marked as
non-final in #3221 to accommodate access to `alterFilters` method for
testing. However, all its other methods can be marked as `final`.
That will avoid warning when ctor uses any of its methods,
like in #3307.

Modifications:

- Mark all `DefaultHttpServerBuilder` methods as `final` except
`alterFilters` that we use in tests.

Result:

No IntelliJ IDEA inspection warnings if `socketOption` method is used in
constructor.