Propagate HTTP/2 streamId to request context
Motivation:

Multiplexed protocols, like HTTP/2 and HTTP/3, use virtual streams for
each request. The streamId is unknown upfront and is assigned by the
protocol at the time client sends meta-data (headers). Currently, there
is no way for client-side to know what was the streamId assigned for a
request.

Modifications:

- Introduce `HttpContextKeys.STREAM_ID` key as a way to query streamId
value from request context, and adjust H2 codec to do so.
- Add tests to validate new behavior.

Result:

`STREAM_ID` value is assigned for a request context as soon as request
meta-data is sent. Users can extract it and correlate with other logs
and events.