fix(grpc-utils): don't log null throwable
#### Motivation

Some logging frameworks get confused and warn when you pass null as the
last argument to a log statement because they can't figure out if its
meant to be a throwable or one of the values you are trying to put into
a placeholder.

#### Modifications

This fixes the LoggingGrpcLifecycleObserver to avoid passing a null
throwable as when no error is present in the request processing chain.

#### Result

No more errors from logging framework for the common case of a
successful response.