refactor(http-utils): make http logging observer consistent with grpc
#### Motivation

Both LoggingHttpLifecycleObserver and LoggingGrpcLifecycleObserver
do the same thing but for different protocols. It would be good to keep
the code consistent between them to ensure they have a similar look and
feel.

#### Modifications

This updates the code in the LoggingHttpLifecycleObserver to match the
code that is in the LoggingGrpcLifecycleObserver

#### Result

Both LifecycleObservers are consistent.