Propagate HTTP/2 streamId to request context (#3311)
#### Motivation

Multiplexed protocols, such as HTTP/2 and HTTP/3, utilize virtual streams for each request. The stream ID is unknown upfront and is assigned by the protocol at the time client sends metadata (headers). Currently, there is no way for client-side to know what the stream ID assigned for a request was.

#### Modifications

- Introduce `HttpContextKeys.STREAM_ID` key as a way to query stream ID value from request context, and adjust H2 codec to do so.
- Add tests to validate new behavior.

#### Result

`STREAM_ID` value is assigned for a request context as soon as request metadata is sent. Users can extract it and correlate with other logs and events. The ID is updated on every retry of the same request instance.