grpc-utils: don't log null throwable  (#3316)
#### Motivation

Some logging frameworks get confused and warn when you pass null as the
last argument to a log statement because they can't figure out if its
meant to be a throwable or one of the values you are trying to put into
a placeholder.

#### Modifications

- Fix the LoggingGrpcLifecycleObserver to avoid passing a null
  throwable as when no error is present in the request processing chain.
- The LoggingHttpLifecycleObserver was also made consistent.

#### Result

No more errors from logging framework for the common case of a
successful response.