loadbalancer: support maintaining a minimum connection pool size per host (#3303)
#### Motivation
Relying on the request path to always establish connections can
be a drag on tail latencies.

#### Modifications
Add an option to keep the hosts connection pool 'warm' by always
having a minimum number of connections. We also have a first-pass
implementation that we can improve upon in the future.

#### Result
The option to get lower tail latencies at the cost of more connection
establishment.