loadbalancer: reduce unnecessary null validation
 #### Motivation

We do an awful lot of `requireNonNull(..)` in the loadbalancer package.
Most of the values we're validating are not user input and our codebase
already enforces the @Nullable annotations, so many of them can be
dropped.

 #### Modifications

Remove a lot of null validations. We also add a few in some places where
we might receive nulls from method calls, eg we are not directly
checking a reference.