opentelemetry-http: make filters stackable (#3324)
#### Motivation

Users may want to add the the OTEL requester filter in multiple positions
in order to build a more detailed view of the trace. Specifically, they may
want a higher level 'logical' filter that encompasses the request call from
the app and then also add the filter as a connection filter so they can track
each physical request sent over the wire.

#### Modifications

Change suppression in the requester filter to check whether the request
should start and then store the result in the context so it can then be
used by a second instance of the filter to make the same decision.

#### Results

Filters can be added multiple times.