opentelemetry-http: append 'Physical ' to physical span names
 #### Motivation

If users are inserting the OTEL requester filter as both a client
filter and a connection filter it can be kind of confusing since
both spans will look identical.

 #### Modifications

Prepend 'Physical ' to the connection level spans if we determine
that there is a parent logical filter in place.

 #### Result

Clearer traces.