opentelemetry-http: Properly extract host and port from authority header (#3336)
 #### Motivation

We currently use the `HostAndPort.ofIpPort` in the GrpcClientAttributesExtractor
to attempt to get our server name and port from the ':authority' pseudo header.
However, this blows up for non-ip names.
Furthermore, because this is scoped to the GrpcClient, we don't get this
benefit for HTTP/2 more generally.

 #### Modifications

Give the HttpAttributesExtractor an expanded notion of 'effectiveHostAndPort'
and do a better job of parsing the :authority header. This is now shared
across the HTTP extraction and the GRPC extraction.