http-netty: use DefaultLoadBalancer API's
 #### Motivation

We have been using the RoundRobinLoadBalancers API's for a
while which give us the ability to switch the default back
to RoundRobinLoadBalancer if we have an emergency. DefaultLoadBalancer
has been the default now for quite a while so lets cut over
to the new API's and get off the deprecated ones.

 #### Modifications

Switch to the new API's.

 #### Result

Less deprecated usages.