loadbalancer: make address assertion in LoadBalancerTest order-independent (#3357)
#### Motivation
A flaky test was identified in `LoadBalancerTest`, specifically `resubscribeToEventsWhenAllHostsAreUnhealthy`, using [NonDex](https://github.com/TestingResearchIllinois/NonDex).
The failure was caused by the use of `assertAddresses(...)`, which relied on a specific order of elements in `lb.usedAddresses()`.
Since the load balancer does not guarantee address ordering, this assertion was nondeterministic across runs.

#### Modifications
- Make test helpers order independent.
- Avoid usage of `assumeTrue` in `roundRobining()`.