http-netty: Don't attempt to use the `.toString()` representation as the host and port (#3370)
 #### Motivation

In our DefaultSingleAddressHttpClientBuilder we have a notion of
`unresolvedAddressToHost`, which lets users configure how arbitrary
address types can be converted into a string representation of a
host and port. The default conversion will fall back to `.toString()`.
Every Object instance has at `.toString()` and almost none of them
are going to be well formatted such that we can parse them into a
host and port.

 #### Modifications

For the default implementation, don't revert to the `.toString()`
for `setFallbackHostAndPort`.