dns-discovery-netty: Add support for DNS backup requests
 #### Motivation

We want to explore whether we can reduce some tail latencies that
seem to be rooted in DNS request timeouts. DNS is most often used
over UDP, which is an inherently lossy protocol transport. When
a packet is lost we have to wait for timeouts and retries, which
can cause tail latencies.

 #### Modifications

Instead of waiting for a full timeout, introduce a backup request
which will fire after a duration. If the first request is slow
because of packet loss or other unfortunate latency then we may
get a faster result from the second.

This is an experimental feature for now, but we can later enhance
it with adaptive backup request deadlines and token buckets to
make sure we're good citizens to the DNS servers.